python_coding_test/requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
